requirements.txt
flask==3.0.2
flask-cors==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.1
openai==1.12.0
pyyaml==6.0.1
gunicorn==21.2.0 